pet-encyclopedia/crawler/requirements.txt
# 宠物百科全书爬虫系统依赖
# Python 3.10+

# =============================================================================
# 核心爬虫框架
# =============================================================================
scrapy>=2.11.0
playwright>=1.40.0
scrapy-playwright>=0.0.33

# =============================================================================
# HTML 解析和内容提取
# =============================================================================
readability-lxml>=0.8.1
trafilatura>=1.6.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
lxml>=5.1.0

# =============================================================================
# Markdown 转换
# =============================================================================
markdownify>=0.11.0

# =============================================================================
# 数据库
# =============================================================================
redis>=5.0.0
pymongo>=4.6.0
psycopg2-binary>=2.9.0

# =============================================================================
# 对象存储
# =============================================================================
minio>=7.2.0

# =============================================================================
# 监控和指标
# =============================================================================
prometheus-client>=0.19.0

# =============================================================================
# 任务队列
# =============================================================================
celery>=5.3.0
rq>=1.15.0

# =============================================================================
# 图片处理
# =============================================================================
pillow>=10.0.0

# =============================================================================
# 去重算法
# =============================================================================
simhash>=2.1.0

# =============================================================================
# 工具库
# =============================================================================
requests>=2.31.0
aiohttp>=3.9.0
pyyaml>=6.0
python-dotenv>=1.0.0

# =============================================================================
# 测试
# =============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# =============================================================================
# 开发工具
# =============================================================================
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
